Move Styles component functionality to StyleList
