+++ react/components/EditorContainer/StoreEditor/Styles/StyleList/StyleCard/Colors.tsx
