+++ react/components/EditorContainer/StoreEditor/Styles/StyleList/StyleCard/index.tsx
