+++ react/components/EditorContainer/StoreEditor/Styles/StyleList/icons/CreateNewIcon.tsx
