+++ react/components/EditorContainer/StoreEditor/Styles/StyleList/queries/CreateStyle.graphql
