+++ react/components/EditorContainer/StoreEditor/Styles/StyleList/queries/DeleteStyle.graphql
