+++ react/components/EditorContainer/StoreEditor/Styles/StyleList/queries/ListStyles.graphql
