+++ react/components/EditorContainer/StoreEditor/Styles/StyleList/queries/SaveSelectedStyle.graphql
